backend/auth-service/requirements.txt
flask==3.0.2
oracledb
python-dotenv
requests==2.32.3
gunicorn==23.0.0